Chore / Move Clinical Service Types (#1185)
* Resolve circular import errors

* Update Imports

* Remove unused variable declaration

* Import fixes; updated switch syntax + comment
