+++ src/common-model/entities.ts
@@ -19,7 +19,6 @@
 
 import { z as zod } from 'zod';
 import { entities as dictionaryEntities } from '@overturebio-stack/lectern-client';
-import { Values } from '../utils/objectTypes';
 
 // this is temporary to keep code compiling until surgery is ready in dictionary, to be removed in favor of
 // the surgery in ClinicalEntitySchemaNames
@@ -94,14 +93,6 @@ export interface ClinicalErrorsResponseRecord {
 	entityName: ClinicalEntitySchemaNames;
 	errors: ClinicalEntityErrorRecord[];
 }
-
-export const ClinicalDataSortTypes = {
-	defaultDonor: 'defaultDonor',
-	invalidEntity: 'invalidEntity',
-	columnSort: 'columnSort',
-};
-
-export type ClinicalDataSortType = Values<typeof ClinicalDataSortTypes>;
 
 export type ClinicalFields =
 	| DonorFieldsEnum
